Created controller class, trying to implement calculate food

+++ src/Zoo/Animal/Diet/Carnivore/Carnivore.cpp
@@ -2,10 +2,10 @@
 
 Carnivore::Carnivore(int _weight, double _ratio) : Diet(_weight, _ratio) {}
 
-int Carnivore::calculateTotalMeat() const {
+double Carnivore::calculateTotalMeat() const {
     return weight*ratio;
 }
 
-int Carnivore::calculateTotalVegetable() const {
+double Carnivore::calculateTotalVegetable() const {
     return 0;
 }
