+++ src/Zoo/Animal/Diet/Herbivore/Herbivore.cpp
@@ -2,10 +2,10 @@
 
 Herbivore::Herbivore(int _weight, double _ratio) : Diet(_weight, _ratio) {}
 
-int Herbivore::calculateTotalMeat() const {
+double Herbivore::calculateTotalMeat() const {
     return 0;
 }
 
-int Herbivore::calculateTotalVegetable() const {
+double Herbivore::calculateTotalVegetable() const {
     return weight*ratio;
 }
